+++ changelog.md
@@ -1,3 +1,3 @@
-## Bugfix
+## General
 
-- Fixed wrong translation in the integrated server without fabric-api
+- Added brief movingBlocks logger